move useOrderbook login into useHydrateStore to reduce rerenders
