+++ hooks/useHydrateStore.tsx
@@ -2,19 +2,43 @@ import { useEffect } from 'react'
 import { AccountInfo } from '@solana/web3.js'
 import useMangoStore from '../stores/useMangoStore'
 import useInterval from './useInterval'
-import useOrderbook from './useOrderbook'
+import { Orderbook as SpotOrderBook, Market } from '@project-serum/serum'
+import {
+  BookSide,
+  BookSideLayout,
+  PerpMarket,
+} from '@blockworks-foundation/mango-client'
 
 const SECONDS = 1000
 const _SLOW_REFRESH_INTERVAL = 20 * SECONDS
+
+function decodeBook(market, accInfo: AccountInfo<Buffer>): number[][] {
+  if (market && accInfo?.data) {
+    const depth = 40
+    if (market instanceof Market) {
+      const book = SpotOrderBook.decode(market, accInfo.data)
+      return book.getL2(depth).map(([price, size]) => [price, size])
+    } else if (market instanceof PerpMarket) {
+      const book = new BookSide(
+        null,
+        market,
+        BookSideLayout.decode(accInfo.data)
+      )
+      return book.getL2(depth).map(([price, size]) => [price, size])
+    }
+  } else {
+    return []
+  }
+}
 
 const useHydrateStore = () => {
   const setMangoStore = useMangoStore((s) => s.set)
   const actions = useMangoStore((s) => s.actions)
   const markets = useMangoStore((s) => s.selectedMangoGroup.markets)
+  const market = useMangoStore((state) => state.selectedMarket.current)
   const marketConfig = useMangoStore((s) => s.selectedMarket.config)
   const selectedMarket = useMangoStore((s) => s.selectedMarket.current)
   const connection = useMangoStore((s) => s.connection.current)
-  useOrderbook()
 
   useEffect(() => {
     actions.fetchMangoGroup()
@@ -38,7 +62,7 @@ const useHydrateStore = () => {
   useEffect(() => {
     let previousBidInfo: AccountInfo<Buffer> | null = null
     let previousAskInfo: AccountInfo<Buffer> | null = null
-    if (!marketConfig) return
+    if (!marketConfig || !market) return
 
     const bidSubscriptionId = connection.onAccountChange(
       marketConfig.bidsKey,
@@ -52,8 +76,8 @@ const useHydrateStore = () => {
         ) {
           previousBidInfo = info
           setMangoStore((state) => {
-            state.accountInfos[marketConfig.bidsKey.toString()] =
-              previousBidInfo
+            state.accountInfos[marketConfig.bidsKey.toString()] = info
+            state.selectedMarket.orderBook.bids = decodeBook(market, info)
           })
         }
       }
@@ -70,8 +94,8 @@ const useHydrateStore = () => {
         ) {
           previousAskInfo = info
           setMangoStore((state) => {
-            state.accountInfos[marketConfig.asksKey.toString()] =
-              previousAskInfo
+            state.accountInfos[marketConfig.asksKey.toString()] = info
+            state.selectedMarket.orderBook.asks = decodeBook(market, info)
           })
         }
       }
@@ -81,7 +105,7 @@ const useHydrateStore = () => {
       connection.removeAccountChangeListener(bidSubscriptionId)
       connection.removeAccountChangeListener(askSubscriptionId)
     }
-  }, [marketConfig, connection, setMangoStore])
+  }, [marketConfig, market, connection, setMangoStore])
 
   // fetch filled trades for selected market
   useInterval(() => {
